additional check for errors
elasticsearch would return a HTTP 200, even if the bulk request failed.
This is a first patch to mitigate the "losing documents" issue
(https://github.com/miku/esbulk/issues/5).

We abort the indexing, if we encounter a single failed response and
suggest the user to lower the number of workers (-w) or to increase
thread_pool.bulk.queue_size in the elasticsearch configuration.

Future versions of esbulk might transparently retry a failed operation
some number of times, but that requires a detailed inspection of the error
response and is therefore postponed from this fix.